CReduce implementation
