+++ include/nbl/builtin/hlsl/scan/declarations.hlsl
@@ -9,8 +9,8 @@
 
 #include "nbl/builtin/hlsl/cpp_compat.hlsl"
 
-#ifndef NBL_BUILTIN_MAX_SCAN_LEVELS
-#define NBL_BUILTIN_MAX_SCAN_LEVELS 7
+#ifndef NBL_BUILTIN_MAX_LEVELS
+#define NBL_BUILTIN_MAX_LEVELS 7
 #endif
 
 namespace nbl
@@ -21,18 +21,18 @@ namespace scan
 {
     // REVIEW: Putting topLevel second allows better alignment for packing of constant variables, assuming lastElement has length 4. (https://learn.microsoft.com/en-us/windows/win32/direct3dhlsl/dx-graphics-hlsl-packing-rules)
     struct Parameters_t {
-        uint32_t lastElement[NBL_BUILTIN_MAX_SCAN_LEVELS/2+1];
+        uint32_t lastElement[NBL_BUILTIN_MAX_LEVELS/2+1];
         uint32_t topLevel;
-        uint32_t temporaryStorageOffset[NBL_BUILTIN_MAX_SCAN_LEVELS/2];
+        uint32_t temporaryStorageOffset[NBL_BUILTIN_MAX_LEVELS/2];
     };
     
     Parameters_t getParameters();
 
     struct DefaultSchedulerParameters_t
     {
-        uint32_t cumulativeWorkgroupCount[NBL_BUILTIN_MAX_SCAN_LEVELS];
-        uint32_t workgroupFinishFlagsOffset[NBL_BUILTIN_MAX_SCAN_LEVELS];
-        uint32_t lastWorkgroupSetCountForLevel[NBL_BUILTIN_MAX_SCAN_LEVELS];
+        uint32_t cumulativeWorkgroupCount[NBL_BUILTIN_MAX_LEVELS];
+        uint32_t workgroupFinishFlagsOffset[NBL_BUILTIN_MAX_LEVELS];
+        uint32_t lastWorkgroupSetCountForLevel[NBL_BUILTIN_MAX_LEVELS];
 
     };
     
@@ -47,7 +47,7 @@ namespace scan
         NBL_CONST_REF_ARG(uint32_t) pseudoLevel
     );
 
-    template<typename Storage_t>
+    template<typename Storage_t, bool isScan>
     void setData(
         NBL_CONST_REF_ARG(Storage_t) data,
         NBL_CONST_REF_ARG(uint32_t) levelInvocationIndex,
