+++ include/nbl/video/utilities/IUtilities.h
@@ -11,6 +11,7 @@
 #include "nbl/video/alloc/StreamingTransientDataBuffer.h"
 #include "nbl/video/utilities/SIntendedSubmitInfo.h"
 #include "nbl/video/utilities/CPropertyPoolHandler.h"
+#include "nbl/video/utilities/CReduce.h"
 #include "nbl/video/utilities/CScanner.h"
 #include "nbl/video/utilities/CComputeBlit.h"
 
@@ -134,6 +135,7 @@ class NBL_API2 IUtilities : public core::IReferenceCounted
             // TODO: investigate whether we need to clamp against 256u instead of 128u on mobile
 #endif
             const auto scan_workgroup_size = core::max(core::roundDownToPoT(limits.maxWorkgroupSize[0]) >> 1u, 128u);
+            m_reducer = core::make_smart_refctd_ptr<CReduce>(core::smart_refctd_ptr(m_device), scan_workgroup_size);
             m_scanner = core::make_smart_refctd_ptr<CScanner>(core::smart_refctd_ptr(m_device), scan_workgroup_size);
         }
 
@@ -161,6 +163,11 @@ class NBL_API2 IUtilities : public core::IReferenceCounted
             return m_propertyPoolHandler.get();
         }
 #endif
+        virtual CReduce* getDefaultReducer() const
+        {
+            return m_reducer.get();
+        }
+
         virtual CScanner* getDefaultScanner() const
         {
             return m_scanner.get();
@@ -721,6 +728,7 @@ class NBL_API2 IUtilities : public core::IReferenceCounted
 #if 0 // TODO: port
         core::smart_refctd_ptr<CPropertyPoolHandler> m_propertyPoolHandler;
 #endif
+        core::smart_refctd_ptr<CReduce> m_reducer;
         core::smart_refctd_ptr<CScanner> m_scanner;
 };
